+++ testFiles/test_parsergenerator/peg/pegparser_src.py
@@ -1,6 +1,6 @@
 # -*- coding:utf-8 -*-
 
-from tacparser import Parser
+from app.baseparser import Parser
 import re
