+++ testFiles/test_reuse_def/subdef02/subdef02parser.py
@@ -1,7 +1,8 @@
 # -*- coding:utf-8 -*-
 
-from tacparser import Parser
 import re
+
+from app.baseparser import Parser
 
 
 class Subdef02Parser(Parser):
